search frontend: move token definitions out of scanner
