+++ client/shared/src/search/query/completion.ts
@@ -1,8 +1,8 @@
 import * as Monaco from 'monaco-editor'
 import { escapeRegExp, startCase } from 'lodash'
 import { FILTERS, resolveFilter } from './filters'
-import { Token } from './scanner'
-import { toMonacoRange } from './tokens'
+import { Token, toMonacoRange } from './tokens'
+
 import { Omit } from 'utility-types'
 import { Observable } from 'rxjs'
 import { IRepository, IFile, ISymbol, ILanguage, IRepoGroup } from '../../graphql/schema'
