+++ client/shared/src/search/query/diagnostics.test.ts
@@ -1,5 +1,6 @@
 import { getDiagnostics } from './diagnostics'
-import { scanSearchQuery, ScanSuccess, ScanResult, Token } from './scanner'
+import { scanSearchQuery, ScanSuccess, ScanResult } from './scanner'
+import { Token } from './tokens'
 import { SearchPatternType } from '../../graphql-operations'
 
 const toSuccess = (result: ScanResult<Token[]>): Token[] => (result as ScanSuccess<Token[]>).term
