+++ client/shared/src/search/query/diagnostics.ts
@@ -1,6 +1,6 @@
 import * as Monaco from 'monaco-editor'
-import { Token } from './scanner'
-import { toMonacoRange } from './tokens'
+import { Token, toMonacoRange } from './tokens'
+
 import { validateFilter } from './filters'
 import { SearchPatternType } from '../../graphql-operations'
