+++ client/web/src/search/queryTelemetry.tsx
@@ -1,5 +1,6 @@
 import { count } from '../../../shared/src/util/strings'
-import { scanSearchQuery, ScanResult, Token } from '../../../shared/src/search/query/scanner'
+import { scanSearchQuery, ScanResult } from '../../../shared/src/search/query/scanner'
+import { Token } from '../../../shared/src/search/query/tokens'
 import { resolveFilter } from '../../../shared/src/search/query/filters'
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
